+++ src/main/java/lotto/domain/prizelotto/FifthPrizeLotto.java
@@ -1,24 +1,23 @@
-package lotto.prizelotto;
+package lotto.domain.prizelotto;
 
 import java.util.List;
 import lotto.RankPrice;
 import lotto.domain.WinNumbers;
 
-public class FirstPrizeLotto extends PrizeLotto {
-
+public class FifthPrizeLotto extends PrizeLotto {
     @Override
     public int calculatePrize() {
-        return RankPrice.FIRST.getPrice() * count;
+        return RankPrice.FIFTH.getPrice() * count;
     }
 
     @Override
     public boolean isSatisfyRule(int count, List<Integer> lottoNumbers, WinNumbers winNumbers) {
-        return count == 6;
+        return count == 3;
     }
 
     @Override
     public int getRank() {
-        return 1;
+        return 5;
     }
 
     @Override
