+++ src/main/java/lotto/domain/prizelotto/FirstPrizeLotto.java
@@ -1,23 +1,24 @@
-package lotto.prizelotto;
+package lotto.domain.prizelotto;
 
 import java.util.List;
 import lotto.RankPrice;
 import lotto.domain.WinNumbers;
 
-public class FifthPrizeLotto extends PrizeLotto {
+public class FirstPrizeLotto extends PrizeLotto {
+
     @Override
     public int calculatePrize() {
-        return RankPrice.FIFTH.getPrice() * count;
+        return RankPrice.FIRST.getPrice() * count;
     }
 
     @Override
     public boolean isSatisfyRule(int count, List<Integer> lottoNumbers, WinNumbers winNumbers) {
-        return count == 3;
+        return count == 6;
     }
 
     @Override
     public int getRank() {
-        return 5;
+        return 1;
     }
 
     @Override
